[CHORE]: 기존 xib 폴더링 및 코드 정리 - issue: #36
