+++ ITZZA/ITZZA/Screen/Community/XibController/ChatInputView.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SnapKit
 
 class ChatInputView: UIView {
     @IBOutlet weak var textInputField: UITextField!
@@ -30,14 +31,9 @@ class ChatInputView: UIView {
         view.backgroundColor = .clear
         self.addSubview(view)
         
-        view.translatesAutoresizingMaskIntoConstraints = false
-        let constraints = [
-            view.topAnchor.constraint(equalTo: self.topAnchor),
-            view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-            view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            view.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-        ]
-        NSLayoutConstraint.activate(constraints)
+        view.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
     }
     
     func setTextField() {
