+++ ITZZA/ITZZA/Screen/Community/XibController/PostContentTableViewHeader.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SnapKit
 
 class PostContentTableViewHeader: UITableViewHeaderFooterView {
     @IBOutlet weak var headerView: ProfileHeaderView!
@@ -33,14 +34,9 @@ class PostContentTableViewHeader: UITableViewHeaderFooterView {
         guard let view = loadXibView(with: Identifiers.postContentTableViewHeader) else { return }
         self.addSubview(view)
 
-        view.translatesAutoresizingMaskIntoConstraints = false
-        let constraints = [
-            view.topAnchor.constraint(equalTo: self.topAnchor),
-            view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-            view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            view.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-        ]
-        NSLayoutConstraint.activate(constraints)
+        view.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
         
         view.addSubview(scrollView)
     }
