+++ ITZZA/ITZZA/Screen/Community/XibController/ProfileHeaderView.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SnapKit
 
 class ProfileHeaderView: UIView {
     @IBOutlet weak var profileImg: UIImageView!
@@ -27,13 +28,8 @@ class ProfileHeaderView: UIView {
         view.backgroundColor = .clear
         self.addSubview(view)
         
-        view.translatesAutoresizingMaskIntoConstraints = false
-        let constraints = [
-            view.topAnchor.constraint(equalTo: self.topAnchor),
-            view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-            view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            view.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-        ]
-        NSLayoutConstraint.activate(constraints)
+        view.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
     }
 }
