--- ITZZA/ITZZA/Screen/Share/Controllers/tmp.swift
@@ -1,8 +1,0 @@
-//
-//  tmp.swift
-//  ITZZA
-//
-//  Created by InJe Choi on 2022/01/29.
-//
-
-import Foundation
