+++ ITZZA/ITZZA/Screen/Share/Xib/PostContentView.xib
@@ -8,18 +8,18 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <customFonts key="customFonts">
-        <array key="SFPRODISPLAYBOLD.OTF">
-            <string>SFProDisplay-Bold</string>
+        <array key="SpoqaHanSansNeo-Medium.otf">
+            <string>SpoqaHanSansNeo-Medium</string>
         </array>
-        <array key="SFPRODISPLAYREGULAR.OTF">
-            <string>SFProDisplay-Regular</string>
+        <array key="SpoqaHanSansNeo-Regular.otf">
+            <string>SpoqaHanSansNeo-Regular</string>
         </array>
     </customFonts>
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="PostContentView" customModule="ITZZA" customModuleProvider="target">
             <connections>
-                <outlet property="postContent" destination="Itv-q2-N3h" id="fcg-8O-35n"/>
-                <outlet property="postTitle" destination="gLA-yl-A6C" id="uHS-Kf-xpp"/>
+                <outlet property="contents" destination="Itv-q2-N3h" id="fcg-8O-35n"/>
+                <outlet property="title" destination="gLA-yl-A6C" id="uHS-Kf-xpp"/>
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
@@ -28,20 +28,20 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gLA-yl-A6C">
-                    <rect key="frame" x="30" y="10" width="37" height="18"/>
+                    <rect key="frame" x="30" y="10" width="39" height="18"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="18" id="BUy-ZY-A7N"/>
                     </constraints>
-                    <fontDescription key="fontDescription" name="SFProDisplay-Bold" family="SF Pro Display" pointSize="15"/>
+                    <fontDescription key="fontDescription" name="SpoqaHanSansNeo-Medium" family="Spoqa Han Sans Neo" pointSize="15"/>
                     <nil key="textColor"/>
                     <nil key="highlightedColor"/>
                 </label>
                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Itv-q2-N3h">
-                    <rect key="frame" x="30" y="36" width="315" height="46"/>
+                    <rect key="frame" x="30" y="38" width="315" height="44"/>
                     <constraints>
                         <constraint firstAttribute="height" relation="greaterThanOrEqual" constant="25" id="RaG-b2-4NC"/>
                     </constraints>
-                    <fontDescription key="fontDescription" name="SFProDisplay-Regular" family="SF Pro Display" pointSize="15"/>
+                    <fontDescription key="fontDescription" name="SpoqaHanSansNeo-Regular" family="Spoqa Han Sans Neo" pointSize="15"/>
                     <nil key="textColor"/>
                     <nil key="highlightedColor"/>
                 </label>
@@ -52,7 +52,7 @@
                 <constraint firstItem="gLA-yl-A6C" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="10" id="5Mf-G5-vBj"/>
                 <constraint firstItem="gLA-yl-A6C" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="30" id="Gtq-dL-E5U"/>
                 <constraint firstItem="Itv-q2-N3h" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="30" id="Rfp-fE-g7i"/>
-                <constraint firstItem="Itv-q2-N3h" firstAttribute="top" secondItem="gLA-yl-A6C" secondAttribute="bottom" constant="8" symbolic="YES" id="T1L-WJ-qnp"/>
+                <constraint firstItem="Itv-q2-N3h" firstAttribute="top" secondItem="gLA-yl-A6C" secondAttribute="bottom" constant="10" id="T1L-WJ-qnp"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="Itv-q2-N3h" secondAttribute="trailing" constant="30" id="blU-DC-1WD"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="Itv-q2-N3h" secondAttribute="bottom" constant="20" id="dON-Gj-LzC"/>
             </constraints>
