+++ ITZZA/ITZZA/Screen/Share/XibControllers/PostContentView.swift
@@ -1,0 +1,45 @@
+//
+//  PostContentView.swift
+//  ITZZA
+//
+//  Created by 황윤경 on 2022/02/06.
+//
+
+import UIKit
+import SnapKit
+
+class PostContentView: UIView {
+    @IBOutlet weak var title: UILabel!
+    @IBOutlet weak var contents: UILabel!
+    
+    override init(frame: CGRect) {
+      super.init(frame: frame)
+        setContentView()
+        configureFont()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        setContentView()
+        configureFont()
+    }
+    
+    private func setContentView() {
+        guard let view = loadXibView(with: Identifiers.postContentView) else { return }
+        view.backgroundColor = .clear
+        self.addSubview(view)
+        
+        view.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+    
+    private func configureFont() {
+        title.textColor = .darkGray6
+        title.font = UIFont.SpoqaHanSansNeoMedium(size: 15)
+        
+        contents.setLineBreakMode()
+        contents.textColor = .darkGray3
+        contents.font = UIFont.SpoqaHanSansNeoRegular(size: 15)
+    }
+}
